* REFACTORING : Customer domain, service, application 수정
