+++ src/main/java/com/io/hhplus/concert/domain/customer/service/CustomerService.java
@@ -3,15 +3,10 @@ package com.io.hhplus.concert.domain.customer.service;
 import com.io.hhplus.concert.application.customer.dto.ChargeCustomerPointServiceResponse;
 import com.io.hhplus.concert.application.customer.dto.CustomerPointBalanceServiceResponse;
 import com.io.hhplus.concert.application.customer.dto.UseCustomerPointServiceResponse;
-import com.io.hhplus.concert.common.enums.PointType;
 import com.io.hhplus.concert.common.enums.ResponseMessage;
 import com.io.hhplus.concert.common.exceptions.CustomException;
-import com.io.hhplus.concert.domain.customer.dto.ChargeCustomerPointServiceRequest;
-import com.io.hhplus.concert.domain.customer.dto.UseCustomerPointServiceRequest;
 import com.io.hhplus.concert.domain.customer.model.Customer;
 import com.io.hhplus.concert.domain.customer.model.CustomerPointHistory;
-import com.io.hhplus.concert.domain.customer.service.model.CustomerModel;
-import com.io.hhplus.concert.domain.customer.service.model.CustomerPointHistoryModel;
 import com.io.hhplus.concert.domain.customer.repository.CustomerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -48,18 +43,16 @@ public class CustomerService {
      */
     @Transactional
     public ChargeCustomerPointServiceResponse chargeCustomerPoint(Long customerId, BigDecimal chargeAmount){
-        // 고객 조회
         Customer customer = customerRepository.findAvailableCustomer(customerId)
                 .filter(o
                         -> o.isNotDreamed()
                         && o.isNotWithdrawn()
                         && o.isNotDeleted())
                 .orElseThrow(() -> new CustomException(ResponseMessage.CUSTOMER_NOT_FOUND));
-        // 포인트 update
-        Customer updatedCustomer = customerRepository.saveCustomer(customer.chargePoint(chargeAmount));
-        // 포인트 내역 insert
-        CustomerPointHistory customerPointHistory = customerRepository.saveCustomerPointHistory(CustomerPointHistory.chargePointOf(customer, chargeAmount));
-        return ChargeCustomerPointServiceResponse.of(updatedCustomer, customerPointHistory);
+        return ChargeCustomerPointServiceResponse.of(
+                customerRepository.saveCustomer(customer.chargePoint(chargeAmount)),
+                customerRepository.saveCustomerPointHistory(CustomerPointHistory.chargePointOf(customer, chargeAmount))
+        );
     }
 
     /**
@@ -70,17 +63,16 @@ public class CustomerService {
      */
     @Transactional
     public UseCustomerPointServiceResponse useCustomerPoint(Long customerId, BigDecimal useAmount) {
-        // 고객 조회
+        // TODO PESSIMISTIC_WRITE VS PESSIMISTIC_READ(갱신소실 문제 있음!!) 알기
+        // 비관적 락도 row 조회 가능하다! 단 select for update는 row 조회 불가
         Customer customer = customerRepository.findAvailableCustomer(customerId)
                 .filter(o
                         -> o.isNotDreamed()
                         && o.isNotWithdrawn()
                         && o.isNotDeleted())
                 .orElseThrow(() -> new CustomException(ResponseMessage.CUSTOMER_NOT_FOUND));
-        // 포인트 update
-        Customer updatedCustomer = customerRepository.saveCustomer(customer.usePoint(useAmount));
-        // 포인트 내역 insert
-        CustomerPointHistory customerPointHistory = customerRepository.saveCustomerPointHistory(CustomerPointHistory.usePointOf(customer, useAmount));
-        return UseCustomerPointServiceResponse.of(updatedCustomer, customerPointHistory);
+        return UseCustomerPointServiceResponse.of(
+                customerRepository.saveCustomer(customer.usePoint(useAmount)),
+                customerRepository.saveCustomerPointHistory(CustomerPointHistory.usePointOf(customer, useAmount)));
     }
 }
